Rename: Category관련 폴더 category package로 이동
- MainCategory, SubCategory 관련 파일 이동

Resolves: #26
