+++ src/main/java/com/example/shinsekai/category/infrastructure/SubCategoryRepository.java
@@ -1,10 +1,9 @@
-package com.example.shinsekai.product.infrastructure;
+package com.example.shinsekai.category.infrastructure;
 
-import com.example.shinsekai.product.entity.category.SubCategory;
+import com.example.shinsekai.category.entity.SubCategory;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
